Reduced memory overhead by not loading weewar maps to memory at startup. This should fix heroku memory limitations issue.

+++ src/main/java/com/github/weewar/mapviewer/dao/WeewarMapDAO.java
@@ -5,5 +5,5 @@ import com.github.weewar.mapviewer.model.WeewarMap;
 import java.util.Optional;
 
 public interface WeewarMapDAO {
-    Optional<WeewarMap> findByMapId(long mapId);
+    Optional<WeewarMap> findByMapId(int mapId);
 }
